release(heltec-v4): publish v2.8.0.9 calibrated ADC and source-identified profiles
Share calibrated battery acquisition and persisted overrides across early boot
and runtime, invalidate stale filter scales, and guard invalid calibration and
storage transactions. Preserve first-region persistence, PRG callback lifetime,
and radio quiescence fixes in both Standard and Solar Router.

Bind firmware versions, ESP descriptors, filenames and SHA256 manifests to the
source identity. Remove obsolete v2.8.0.8 audit notes, unused helpers, duplicate
includes and temporary arrays; simplify the release description and build helper.

Validation: trunk fmt; nine host gates with sanitizer coverage pass. The preceding
implementation passed 669 native cases and Solar Router hardware calibration,
configuration-preservation, PRG and Bluetooth checks. Final clean-commit builds
and native suites are required before publication; release CI repeats the gates.

+++ src/graphics/Screen.h
@@ -6,6 +6,7 @@
 #include "mesh/generated/meshtastic/config.pb.h"
 #include <OLEDDisplay.h>
 #include <atomic>
+#include <cstddef>
 #include <functional>
 #include <memory>
 #include <string>
@@ -38,9 +39,9 @@ struct BannerOverlayOptions {
     uint32_t durationMs = 30000;
     const char **optionsArrayPtr = nullptr;
     const int *optionsEnumPtr = nullptr;
-    uint8_t optionsCount = 0;
+    size_t optionsCount = 0;
     std::function<void(int)> bannerCallback = nullptr;
-    int8_t InitialSelected = 0;
+    int InitialSelected = 0;
     notificationTypeEnum notificationType = notificationTypeEnum::text_banner;
 };
 } // namespace graphics
